refactor: Change Core directory
