+++ src/packages/core/diff.ts
